Remove workspace filtering from worklist/schedule
