+++ src/js/apps/patients/worklist/filters_app.js
@@ -1,24 +1,17 @@
 import Radio from 'backbone.radio';
 
-import intl from 'js/i18n';
-
 import App from 'js/base/app';
 
 import StateModel from 'js/apps/patients/sidebar/filters_state';
 
 import AllFiltersApp from 'js/apps/patients/sidebar/filters-sidebar_app';
 
-import { FiltersView, WorkspacesDropList, AllFiltersButtonView } from 'js/views/patients/worklist/filters_views';
-
-const i18n = intl.patients.worklist.filtersApp;
+import { AllFiltersButtonView } from 'js/views/patients/worklist/worklist_views';
 
 export default App.extend({
   StateModel,
   childApps: {
     allFilters: AllFiltersApp,
-  },
-  stateEvents: {
-    'change:filters': 'showWorkspacesFilterView',
   },
   onBeforeStop() {
     this.toggleFiltering(false);
@@ -26,26 +19,17 @@ export default App.extend({
   onStart({ availableStates }) {
     const currentClinician = Radio.request('bootstrap', 'currentUser');
     this.canViewAssignedActions = currentClinician.can('app:worklist:clinician_filter');
-    this.workspaces = currentClinician.getWorkspaces();
     this.availableStates = availableStates;
 
-    this.showView(new FiltersView());
-    this.showFilters();
+    const allFiltersButtonView = this.showView(new AllFiltersButtonView({
+      model: this.getState(),
+    }));
+
+    this.listenTo(allFiltersButtonView, 'click', this.showFiltersSidebar);
   },
   toggleFiltering(isFiltering) {
     this.isFiltering = isFiltering;
     this.trigger('toggle:filtersSidebar', isFiltering);
-  },
-  showFilters() {
-    this.showAllFiltersButtonView();
-    this.showWorkspacesFilterView();
-  },
-  showAllFiltersButtonView() {
-    const ownerView = this.showChildView('allFilters', new AllFiltersButtonView({
-      model: this.getState(),
-    }));
-
-    this.listenTo(ownerView, 'click', this.showFiltersSidebar);
   },
   showFiltersSidebar() {
     if (this.isFiltering) return;
@@ -72,32 +56,5 @@ export default App.extend({
     this.listenTo(sidebar, 'stop', () => {
       this.toggleFiltering(false);
     });
-  },
-  showWorkspacesFilterView() {
-    if (this.workspaces.length < 2) return;
-
-    const workspaces = this._getWorkspaces();
-    const selected = workspaces.get(this.getState().getFilter('workspaceId')) || workspaces.at(0);
-
-    const workspacesFilter = new WorkspacesDropList({
-      collection: workspaces,
-      state: { selected },
-    });
-
-    this.listenTo(workspacesFilter.getState(), 'change:selected', (state, { id }) => {
-      this.getState().setFilter('workspaceId', id);
-    });
-
-    this.showChildView('workspace', workspacesFilter);
-  },
-  _getWorkspaces() {
-    const workspaces = this.workspaces.clone();
-
-    workspaces.unshift({
-      id: null,
-      name: i18n.allWorkspacesName,
-    });
-
-    return workspaces;
   },
 });
