+++ src/js/views/patients/schedule/schedule-list.scss
@@ -1,3 +1,7 @@
+.schedule__list-search {
+  margin-left: auto;
+}
+
 .schedule-list__header {
   border-bottom: 1px solid $black-90;
   color: $black-60;
