--- src/js/views/patients/schedule/schedule.scss
@@ -1,17 +1,0 @@
-.schedule__filters {
-  align-items: center;
-  display: flex;
-  width: 100%;
-}
-
-.schedule__all-filters:not(:empty) {
-  margin-right: 24px;
-}
-
-.schedule__filter:not(:empty) {
-  margin-right: 8px;
-}
-
-.schedule__list-search {
-  margin-left: auto;
-}
