+++ src/js/views/patients/sidebar/filters/filters-sidebar_views.js
@@ -15,8 +15,6 @@ import './filters-sidebar.scss';
 
 const i18n = intl.patients.sidebar.filters.filtersSidebarViews;
 
-const workspaceLabelView = i18n.workspaceLabelView;
-
 const CustomFilterDropList = Droplist.extend({
   popWidth() {
     return this.getView().$el.outerWidth();
@@ -226,5 +224,4 @@ export {
   HeaderView,
   CustomFiltersView,
   StatesFiltersView,
-  workspaceLabelView,
 };
